requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==1.11.13
djangorestframework==3.8.2
idna==2.6
oauthlib==2.1.0
PyJWT==1.6.3
python-openid==2.2.5
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
Unidecode==1.0.22
urllib3==1.22
